refactor(shell): extract AliasValue/AliasesError into shell/alias.rs
Re-exported from shell.rs so existing paths are unchanged. Also drops the
low-value factory_tests (which exec'd real shells to assert canonical_name).
