+++ crates/atuin-common/src/shell.rs
@@ -12,6 +12,7 @@ use serde::Serialize;
 use sysinfo::{Process, System, get_current_pid};
 use thiserror::Error;
 
+mod alias;
 mod posix;
 
 pub mod bash;
@@ -19,6 +20,8 @@ pub mod fish;
 pub mod sh;
 pub mod xonsh;
 pub mod zsh;
+
+pub use alias::{AliasValue, AliasesError};
 
 #[derive(Debug, Clone, thiserror::Error)]
 pub enum RunError {
@@ -36,60 +39,6 @@ pub enum RunError {
     },
     #[error("the output of '{command}' was not delimited as expected")]
     Delimiter { command: String },
-}
-
-#[derive(Debug, Clone, thiserror::Error)]
-pub enum AliasesError {
-    #[error("could not query the shell's aliases: {0}")]
-    Run(#[from] RunError),
-    #[error("could not parse the shell's alias list at byte {offset}, near {near:?}")]
-    Parse { offset: usize, near: String },
-    #[error("the alias probe did not run to completion")]
-    Probe,
-}
-
-/// The body an alias expands to.
-///
-/// Shells disagree about what an alias body *is*. Most store a command string
-/// the shell re-parses on use; xonsh stores an argv vector it execs directly.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum AliasValue {
-    /// A command string, re-parsed by the shell: bash, zsh, sh, fish.
-    Command(BString),
-    /// An argv vector, exec'd without a shell parsing pass: xonsh.
-    Argv(Vec<BString>),
-}
-
-impl AliasValue {
-    /// Render the alias body as a command string for a POSIX shell.
-    ///
-    /// `Command` already is one. `Argv` is single-quoted argument by argument so
-    /// a POSIX re-parse reproduces the original argv: a plain space join would
-    /// lose the boundaries of any argument containing a space and drop empty
-    /// arguments entirely.
-    pub fn shcmd(&self) -> BString {
-        match self {
-            AliasValue::Command(cmd) => cmd.clone(),
-            AliasValue::Argv(argv) => {
-                let mut out = BString::default();
-                for (index, arg) in argv.iter().enumerate() {
-                    if index > 0 {
-                        out.push(b' ');
-                    }
-                    out.push(b'\'');
-                    for &byte in arg.iter() {
-                        if byte == b'\'' {
-                            out.extend_from_slice(br"'\''");
-                        } else {
-                            out.push(byte);
-                        }
-                    }
-                    out.push(b'\'');
-                }
-                out
-            }
-        }
-    }
 }
 
 #[async_trait::async_trait]
@@ -280,49 +229,6 @@ impl Shell {
     }
 }
 
-#[cfg(test)]
-mod alias_value_tests {
-    use super::AliasValue;
-    use bstr::BString;
-
-    #[test]
-    fn command_shcmd_is_passthrough() {
-        let v = AliasValue::Command(BString::from("ls -l"));
-        assert_eq!(v.shcmd(), BString::from("ls -l"));
-    }
-
-    #[test]
-    fn argv_shcmd_single_quotes_each_argument() {
-        let v = AliasValue::Argv(vec![
-            BString::from("git"),
-            BString::from("commit"),
-            BString::from("-m"),
-            BString::from("hello world"),
-        ]);
-        assert_eq!(
-            v.shcmd(),
-            BString::from(r"'git' 'commit' '-m' 'hello world'")
-        );
-    }
-
-    #[test]
-    fn argv_shcmd_preserves_empty_and_escapes_quote() {
-        assert_eq!(
-            AliasValue::Argv(vec![
-                BString::from("echo"),
-                BString::from(""),
-                BString::from("x")
-            ])
-            .shcmd(),
-            BString::from(r"'echo' '' 'x'"),
-        );
-        assert_eq!(
-            AliasValue::Argv(vec![BString::from("echo"), BString::from("it's")]).shcmd(),
-            BString::from(r"'echo' 'it'\''s'"),
-        );
-    }
-}
-
 pub fn shell_name(parent: Option<&Process>) -> String {
     let sys = System::new_all();
 
@@ -341,25 +247,4 @@ pub fn shell_name(parent: Option<&Process>) -> String {
     let shell = shell.strip_prefix('-').unwrap_or(&shell);
 
     shell.to_string()
-}
-
-#[cfg(test)]
-mod factory_tests {
-    use super::Shell;
-
-    #[tokio::test]
-    async fn interface_maps_supported_shells_to_the_right_impl() {
-        assert_eq!(Shell::Bash.interface().unwrap().canonical_name(), "bash");
-        assert_eq!(Shell::Sh.interface().unwrap().canonical_name(), "sh");
-        assert_eq!(Shell::Zsh.interface().unwrap().canonical_name(), "zsh");
-        assert_eq!(Shell::Fish.interface().unwrap().canonical_name(), "fish");
-        assert_eq!(Shell::Xonsh.interface().unwrap().canonical_name(), "xonsh");
-    }
-
-    #[test]
-    fn interface_is_none_for_unsupported_shells() {
-        assert!(Shell::Nu.interface().is_none());
-        assert!(Shell::Powershell.interface().is_none());
-        assert!(Shell::Unknown.interface().is_none());
-    }
 }
